+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 + [x] `select` - fixed method `val()` no icons were displayed when using this method
 + [x] `select` - added method `clear()` to clear selected value(s)
 + [x] `select` - refactored method `reset()` - now calling this method resets select to initial state
++ [x] `typer` - fixed color change